Typo fixes and extensive changes (see extdesc)
Images are no longer required to have a specified height in this release because a maximum height of 20 pixels is established in the beginning of the code. While this allows for some slight imperfections where the moue cursor will be hovering on nothing, it's only so many pixels and the end user won't notice unless said user tries to actively find the empty space around an image.

If it works out like I think, most names will appear perfectly centered.